Since now on Indexer result contains number of found matches for each image.

+++ src/main/algorithm/indexer/cross_indexer/CrossIndexer.h
@@ -52,7 +52,7 @@ namespace feitir {
          * @param expansionSize
          * @return
          */
-        std::vector<ImageBSIFT::BSIFT> expandCodeWord(const ImageBSIFT::BSIFT &codeWord, const size_t expansionSize);
+        std::vector<ImageBSIFT::BSIFT> expandCodeWord(const ImageBSIFT::BSIFT &codeWord, const int expansionSize);
 
     protected:
         /**
